+++ src/libxtp/gwbse/gw.cc
@@ -1,5 +1,5 @@
 /*
- *            Copyright 2009-2020 The VOTCA Development Team
+ *            Copyright 2009-2021 The VOTCA Development Team
  *                       (http://www.votca.org)
  *
  *      Licensed under the Apache License, Version 2.0 (the "License")
@@ -27,24 +27,17 @@
 #include "votca/xtp/gw.h"
 #include "votca/xtp/newton_rapson.h"
 #include "votca/xtp/rpa.h"
-#include "votca/xtp/sigma_cda.h"
-#include "votca/xtp/sigma_exact.h"
-#include "votca/xtp/sigma_ppm.h"
+#include "votca/xtp/sigmafactory.h"
 
 namespace votca {
 namespace xtp {
 
 void GW::configure(const options& opt) {
+  Sigma().RegisterAll();
   opt_ = opt;
   qptotal_ = opt_.qpmax - opt_.qpmin + 1;
   rpa_.configure(opt_.homo, opt_.rpamin, opt_.rpamax);
-  if (opt_.sigma_integration == "exact") {
-    sigma_ = std::make_unique<Sigma_Exact>(Mmn_, rpa_);
-  } else if (opt_.sigma_integration == "cda") {
-    sigma_ = std::make_unique<Sigma_CDA>(Mmn_, rpa_);
-  } else if (opt_.sigma_integration == "ppm") {
-    sigma_ = std::make_unique<Sigma_PPM>(Mmn_, rpa_);
-  }
+  sigma_ = Sigma().Create(opt_.sigma_integration, Mmn_, rpa_);
   Sigma_base::options sigma_opt;
   sigma_opt.homo = opt_.homo;
   sigma_opt.qpmax = opt_.qpmax;