+++ src/tests/test_sigma_ppm.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright 2009-2020 The VOTCA Development Team (http://www.votca.org)
+ * Copyright 2009-2021 The VOTCA Development Team (http://www.votca.org)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,6 +13,7 @@
  * limitations under the License.
  *
  */
+#include "votca/xtp/sigma_base.h"
 #define BOOST_TEST_MAIN
 
 #define BOOST_TEST_MODULE sigma_test
@@ -32,7 +33,7 @@
 #include "votca/xtp/orbitals.h"
 #include "votca/xtp/ppm.h"
 #include "votca/xtp/rpa.h"
-#include "votca/xtp/sigma_ppm.h"
+#include "votca/xtp/sigmafactory.h"
 #include "votca/xtp/threecenter.h"
 #include <libint2/initialize.h>
 using namespace votca::xtp;
@@ -67,18 +68,19 @@ BOOST_AUTO_TEST_CASE(sigma_full) {
   rpa.configure(4, 0, 16);
   rpa.setRPAInputEnergies(mo_energy);
 
-  Sigma_PPM sigma = Sigma_PPM(Mmn, rpa);
+  Sigma().RegisterAll();
+  std::unique_ptr<Sigma_base> sigma = Sigma().Create("ppm", Mmn, rpa);
 
-  Sigma_PPM::options opt;
+  Sigma_base::options opt;
   opt.homo = 4;
   opt.qpmin = 0;
   opt.qpmax = 16;
   opt.rpamin = 0;
   opt.rpamax = 16;
   opt.eta = 1e-3;
-  sigma.configure(opt);
+  sigma->configure(opt);
 
-  Eigen::MatrixXd x = sigma.CalcExchangeMatrix();
+  Eigen::MatrixXd x = sigma->CalcExchangeMatrix();
 
   Eigen::MatrixXd x_ref = votca::tools::EigenIO_MatrixMarket::ReadMatrix(
       std::string(XTP_TEST_DATA_FOLDER) + "/sigma_ppm/x_ref.mm");
@@ -92,9 +94,9 @@ BOOST_AUTO_TEST_CASE(sigma_full) {
   }
   BOOST_CHECK_EQUAL(check_x, true);
 
-  sigma.PrepareScreening();
-  Eigen::MatrixXd c = sigma.CalcCorrelationOffDiag(mo_energy);
-  c.diagonal() = sigma.CalcCorrelationDiag(mo_energy);
+  sigma->PrepareScreening();
+  Eigen::MatrixXd c = sigma->CalcCorrelationOffDiag(mo_energy);
+  c.diagonal() = sigma->CalcCorrelationDiag(mo_energy);
 
   Eigen::MatrixXd c_ref = votca::tools::EigenIO_MatrixMarket::ReadMatrix(
       std::string(XTP_TEST_DATA_FOLDER) + "/sigma_ppm/c_ref.mm");