+++ CHANGELOG.rst
@@ -47,6 +47,7 @@ Version 2022-dev
 -  overhaul complete option handling (#704)
 -  Fixed exciton options checking (#726)
 -  added basis gpu runner and test to suite (#725)
+-  turned sigma choice into a factory (#731)
 
 Version 2021.2 (released XX.07.21)
 ==================================